+++ core/changelog.md
@@ -5,3 +5,4 @@
 - feat: request path override functionality to support full URLs (with scheme and host) as well as custom paths
 - fix: missing and duplicated tool results in Bedrock - [@hhieuu](https://github.com/hhieuu)
 - fix: support HuggingFace model names without an explicit provider prefix
+- feat: adds support for custom OAuth scopes when authenticating with Azure Entra ID